Your on your way to implementing temperature. It would be a good idea to add An entire "TemperatureInterval" to the Plant model. This could allow you to set the value on the picker without having to go through all this complex logic.

+++ FloraSaver/Models/TemperatureInterval.cs
@@ -15,7 +15,7 @@ namespace FloraSaver.Models
         
         private int _temperatureLevel;
         public int TemperatureLevel {
-            get { return _temperatureLevel; }
+            get { return (int)_temperatureLevel; }
             set
             {
                 _temperatureLevel = value;
