+++ FloraSaver/Services/PickerService.cs
@@ -86,12 +86,12 @@ namespace FloraSaver.Services
         {
             return new List<TemperatureInterval>
             {
-                new TemperatureInterval(){IntervalText = $"100°: Air conditioning is for the weak!", TemperatureLevel =  100 },
-                new TemperatureInterval(){IntervalText = $"80°-99°: Thriving in my own lane",TemperatureLevel =  80},
-                new TemperatureInterval(){IntervalText = $"60°-79°: Normal Indoor Temperatures",TemperatureLevel =  60 },
-                new TemperatureInterval(){IntervalText = $"40°-59°: We got blankets. We're living good",TemperatureLevel =  40  },
-                new TemperatureInterval(){IntervalText = $"20°-39°: It's only a little frosty",TemperatureLevel =  20 },
-                new TemperatureInterval(){IntervalText = $"0°-19°: They call my room 'fridge space'",TemperatureLevel =  0 }
+                new TemperatureInterval(){IntervalText = $"100°: Air conditioning is for the weak!", TemperatureLevel =  (int)100 },
+                new TemperatureInterval(){IntervalText = $"80°-99°: Thriving in my own lane",TemperatureLevel =  (int)80},
+                new TemperatureInterval(){IntervalText = $"60°-79°: Normal Indoor Temperatures",TemperatureLevel =  (int)60 },
+                new TemperatureInterval(){IntervalText = $"40°-59°: We got blankets. We're living good",TemperatureLevel =  (int)40  },
+                new TemperatureInterval(){IntervalText = $"20°-39°: It's only a little frosty",TemperatureLevel =  (int)20 },
+                new TemperatureInterval(){IntervalText = $"0°-19°: They call my room 'fridge space'",TemperatureLevel =  (int)0 }
             };
         }
 
@@ -99,12 +99,12 @@ namespace FloraSaver.Services
         {
             return new List<TemperatureInterval>
             {
-                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(100)}°+: Air conditioning is for the weak!", TemperatureLevel =  100, IsCelsius = true },
-                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(80)}°-{TemperatureInterval.ConvertToCelsius(99)}°: Thriving in my own lane",TemperatureLevel =  80, IsCelsius = true},
-                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(60)}°-{TemperatureInterval.ConvertToCelsius(79)}°: Normal Indoor Temperatures",TemperatureLevel =  60, IsCelsius = true },
-                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(40)}°-{TemperatureInterval.ConvertToCelsius(59)}°: We got blankets. We're living good",TemperatureLevel =  40, IsCelsius = true  },
-                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(20)}°-{TemperatureInterval.ConvertToCelsius(39)}°: It's only a little frosty",TemperatureLevel =  20, IsCelsius = true },
-                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(0)}°-{TemperatureInterval.ConvertToCelsius(19)}°: They call my room 'fridge space'",TemperatureLevel =  0, IsCelsius = true }
+                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(100)}°+: Air conditioning is for the weak!", TemperatureLevel =  (int)100, IsCelsius = true },
+                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(80)}°-{TemperatureInterval.ConvertToCelsius(99)}°: Thriving in my own lane",TemperatureLevel =  (int)80, IsCelsius = true},
+                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(60)}°-{TemperatureInterval.ConvertToCelsius(79)}°: Normal Indoor Temperatures",TemperatureLevel =  (int)60, IsCelsius = true },
+                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(40)}°-{TemperatureInterval.ConvertToCelsius(59)}°: We got blankets. We're living good",TemperatureLevel =  (int)40, IsCelsius = true  },
+                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(20)}°-{TemperatureInterval.ConvertToCelsius(39)}°: It's only a little frosty",TemperatureLevel =  (int)20, IsCelsius = true },
+                new TemperatureInterval(){IntervalText = $"{TemperatureInterval.ConvertToCelsius(0)}°-{TemperatureInterval.ConvertToCelsius(19)}°: They call my room 'fridge space'",TemperatureLevel =  (int)0, IsCelsius = true }
             };
         }
     }
